Rename "dialect" to "preprocessor"
