+++ .prettierignore
@@ -1,1 +1,2 @@
-packages/dialect-cics/src/generated/
+packages/preprocessor-cics/src/generated/
+packages/preprocessor-db2/src/generated/
