+++ packages/language/package.json
@@ -40,9 +40,9 @@
     "@chevrotain/types": "^11.1.1",
     "@chevrotain/utils": "^11.1.1",
     "chevrotain": "^11.1.1",
-    "dialect-api": "workspace:*",
-    "dialect-cics": "workspace:*",
-    "dialect-sql": "workspace:*",
+    "preprocessor-api": "workspace:*",
+    "preprocessor-cics": "workspace:*",
+    "preprocessor-db2": "workspace:*",
     "jsonc-parser": "4.0.0-next.1",
     "lodash-es": "^4.17.23",
     "lru-cache": "^11.2.5",
