+++ packages/preprocessor-cics/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "dialect-sql",
-  "description": "The dialect SQL package",
+  "name": "preprocessor-cics",
+  "description": "The CICS preprocessor package",
   "version": "1.0.0",
   "type": "module",
   "engines": {
@@ -26,16 +26,16 @@
     }
   },
   "scripts": {
-    "clean": "shx rm -fr *.tsbuildinfo out",
+    "clean": "shx rm -fr *.tsbuildinfo out generated",
     "grammar": "pnpm run lexer && pnpm run parser",
-    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/Db2SqlExecLexer.g4",
-    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/Db2SqlExecParser.g4"
+    "lexer": "antlr-ng -Dlanguage=TypeScript -l -o src/generated ./src/antlr/CICSLexer.g4",
+    "parser": "antlr-ng -Dlanguage=TypeScript -v -o src/generated ./src/antlr/CICSParser.g4"
   },
   "devDependencies": {
     "antlr-ng": "^1.0.10"
   },
   "dependencies": {
-    "dialect-api": "workspace:*",
+    "preprocessor-api": "workspace:*",
     "antlr4ng": "^3.0.16",
     "vscode-languageserver": "~9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
