+++ packages/preprocessor-cics/src/engine/parsing.ts
@@ -21,7 +21,6 @@
  * Contributors:
  *   Broadcom, Inc. - initial API and implementation
  */
-
 import {
   BaseErrorListener,
   Recognizer,
@@ -29,9 +28,9 @@ import {
   RecognitionException,
   Token as AntlrToken,
 } from "antlr4ng";
-import { Db2SqlExecParserVisitor } from "../generated/Db2SqlExecParserVisitor";
-import { Dbs_sql_identifierContext } from "../generated/Db2SqlExecParser";
-import { ParseError, SemanticsKind, Token } from "dialect-api";
+import { CICSParserVisitor } from "../generated/CICSParserVisitor";
+import { CicsWordContext } from "../generated/CICSParser";
+import { ParseError, SemanticsKind, Token } from "preprocessor-api";
 
 export class CollectingErrorListener extends BaseErrorListener {
   public readonly errors: ParseError[] = [];
@@ -39,27 +38,28 @@ export class CollectingErrorListener extends BaseErrorListener {
   override syntaxError<S extends AntlrToken, T extends ATNSimulator>(
     _recognizer: Recognizer<T>,
     offendingSymbol: S | null,
-    _line: number,
-    _charPositionInLine: number,
+    line: number,
+    charPositionInLine: number,
     msg: string,
     _e: RecognitionException | null,
   ): void {
     this.errors.push({
       message: msg,
-      startOffset: offendingSymbol?.start ?? 0,
-      endOffset: offendingSymbol?.stop ?? 0,
+      startOffset: offendingSymbol?.start || 0,
+      endOffset: offendingSymbol?.stop || 0,
     });
   }
 }
 
-export class CollectingIdentifierVisitor extends Db2SqlExecParserVisitor<void> {
+export class CollectingIdentifierVisitor extends CICSParserVisitor<void> {
   readonly identifiers: Token[] = [];
-  override visitDbs_sql_identifier = (ctx: Dbs_sql_identifierContext): void => {
-    if (ctx.start && ctx.stop) {
+  override visitCicsWord = (ctx: CicsWordContext): void => {
+    const symbol = ctx.WORD_IDENTIFIER()?.getSymbol();
+    if (symbol && symbol.text) {
       this.identifiers.push({
-        image: ctx.getText(),
-        startOffset: ctx.start.start,
-        endOffset: ctx.stop.stop,
+        image: symbol.text,
+        startOffset: symbol.start,
+        endOffset: symbol.stop,
         semanticsKind: SemanticsKind.Identifier,
       });
     }
